+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Removed
 
+- Removed `phylum group transfer` subcommand
 - Removed owner email from `phylum group list` results
 
 ## 6.1.2 - 2024-02-22